create post, put/patch, delete apis and its ResponseEntities.

+++ src/main/java/com/storeapi/store/dtos/ProductDto.java
@@ -1,9 +1,6 @@
 package com.storeapi.store.dtos;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.Getter;
-
 import java.math.BigDecimal;
 
 @Data
@@ -12,5 +9,5 @@ public class ProductDto {
     private String name;
     private BigDecimal price;
     private String description;
-    private Long categoryId;
+    private Byte categoryId;
 }
